+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Optimization: Reduce the group size of the deferred lighting pass from 16x16 to 8x8
 - Replaced HDCamera.computePassCount by viewCount
+- Removed xrInstancing flag in RTHandles (replaced by TextureXR.slices and TextureXR.dimensions) 
 
 ## [6.7.0-preview] - 2019-05-16